Spider Quotes Scraper 2 v 1.3.0.3

+++ quotes_scraper/quotes_scraper/quotes_scraper/spiders/quotes_2.py
@@ -17,9 +17,10 @@ class quotesSpider(scrapy.Spider):
     start_urls = [
         'http://quotes.toscrape.com/'
     ]
-    
-    os.remove("quotes_2.csv")
 
+    if os.path.exists('quotes_2.csv'):
+        os.remove("quotes_2.csv")
+        
     custom_settings = {
         "FEEDS":{"quotes_2.csv":{"format":"csv"}}
     }
